A bit of cleanup and renaming to make room for AndHowTestInit usage in the andhow-multimodule-dataprocess test
