+++ andhow-simulated-app-tests/andhow-multimodule-dataprocess/andhow-default-behavior-dep1/pom.xml
@@ -2,27 +2,25 @@
 	<modelVersion>4.0.0</modelVersion>
 	<parent>
 		<groupId>org.yarnandtail</groupId>
-		<artifactId>andhow-parent</artifactId>
+		<artifactId>andhow-multimodule-dataprocess</artifactId>
 		<version>0.4.0-SNAPSHOT</version>
 	</parent>
 
-	<artifactId>andhow-default-behavior-dependency</artifactId>
+	<artifactId>andhow-default-behavior-dep1</artifactId>
 	
 	<!-- This really isn't meant to be packaged up at all - its sample code -->
 	<packaging>jar</packaging>
-	<name>andhow-default-behavior-dependency</name>
-	<description>Sample to be used as a compiled jar by the andhow-default-behavior-test to test property behavior in external compiled libraries.</description>
-
+	<name>andhow-default-behavior-dep1</name>
+	<description>Sample to be used as a compiled jar by the 
+		andhow-default-behavior-test to test property behavior
+		in external compiled libraries.</description>
+		
 	<dependencies>
 		<dependency>
+			<!-- Using the combined, single jar dependency -->
 			<groupId>org.yarnandtail</groupId>
-			<artifactId>andhow-core</artifactId>
-			<version>${project.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>org.yarnandtail</groupId>
-			<artifactId>andhow-annotation-processor</artifactId>
-			<version>${project.version}</version>
+			<artifactId>andhow</artifactId>
+			<version>${andhow-version-to-use}</version>
 		</dependency>
 		<dependency>
 			<groupId>junit</groupId>
