+++ andhow-simulated-app-tests/andhow-multimodule-dataprocess/andhow-default-behavior-dep1/src/main/java/com/dep1/EarthMapMaker.java
@@ -1,4 +1,4 @@
-package com.map;
+package com.dep1;
 
 import org.yarnandtail.andhow.*;
 import org.yarnandtail.andhow.property.*;
