+++ andhow-simulated-app-tests/andhow-multimodule-dataprocess/andhow-default-behavior-test/pom.xml
@@ -2,7 +2,7 @@
 	<modelVersion>4.0.0</modelVersion>
 	<parent>
 		<groupId>org.yarnandtail</groupId>
-		<artifactId>andhow-parent</artifactId>
+		<artifactId>andhow-multimodule-dataprocess</artifactId>
 		<version>0.4.0-SNAPSHOT</version>
 	</parent>
 
@@ -16,24 +16,20 @@
 	<dependencies>
 		<dependency>
 			<groupId>org.yarnandtail</groupId>
-			<artifactId>andhow-core</artifactId>
+			<artifactId>andhow-default-behavior-dep1</artifactId>
 			<version>${project.version}</version>
 		</dependency>
 		<dependency>
+			<!-- Using the combined, single jar dependency -->
 			<groupId>org.yarnandtail</groupId>
-			<artifactId>andhow-annotation-processor</artifactId>
-			<version>${project.version}</version>
+			<artifactId>andhow</artifactId>
+			<version>${andhow-version-to-use}</version>
 		</dependency>
 		<dependency>
 			<groupId>${project.groupId}</groupId>
 			<artifactId>andhow-test-harness</artifactId>
 			<version>${project.version}</version>
 			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.yarnandtail</groupId>
-			<artifactId>andhow-default-behavior-dependency</artifactId>
-			<version>${project.version}</version>
 		</dependency>
 		<dependency>
 			<groupId>junit</groupId>
