+++ andhow-simulated-app-tests/andhow-multimodule-dataprocess/andhow-default-behavior-test/src/main/resources/andhow.properties
@@ -21,28 +21,28 @@ org.dataprocess.ExternalServiceConnector.ConnectionConfig.TIMEOUT = 60
 
 # ##########################################################################################
 # Property Group 'Configuration for map generation' - Bounds define max map extents.
-# 	Defined in interface com.map.EarthMapMaker
+# 	Defined in interface com.dep1.EarthMapMaker
 
 # 
 # EAST_BOUND (Integer) NON-NULL - East-most edge of map, in deg. longitude
 # Default Value: -66
-com.map.EarthMapMaker.EAST_BOUND = -65
+com.dep1.EarthMapMaker.EAST_BOUND = -65
 
 # 
 # MAP_NAME (String) NON-NULL - Name displayed at the top of the map
-com.map.EarthMapMaker.MAP_NAME = My Map
+com.dep1.EarthMapMaker.MAP_NAME = My Map
 
 # 
 # NORTH_BOUND (Integer) NON-NULL - North-most edge of map, in deg. latitue
 # Default Value: 50
-com.map.EarthMapMaker.NORTH_BOUND = 51
+com.dep1.EarthMapMaker.NORTH_BOUND = 51
 
 # 
 # SOUTH_BOUND (Integer) NON-NULL - South-most edge of map, in deg. latitue
 # Default Value: 24
-com.map.EarthMapMaker.SOUTH_BOUND = 23
+com.dep1.EarthMapMaker.SOUTH_BOUND = 23
 
 # 
 # WEST_BOUND (Integer) NON-NULL - West-most edge of map, in deg. longitude
 # Default Value: -124
-com.map.EarthMapMaker.WEST_BOUND = -125
+com.dep1.EarthMapMaker.WEST_BOUND = -125
